Hack together an initial reproducer for libcrypto dlopen issue
This has been run on FreeBSD 15.0-CURRENT--where it crashes (likely needs
some polishing) and OSX--where it succeeds.

This project uses a combination of bmake and cmake to build the demo exe
and library. I opted for bmake under FreeBSD to improve debugging
introspection, slightly, given that the out of box defaults are a bit
better with respect to WARNS, gdb/lldb, etc.